fix(16): satisfy mypy --strict + bandit pre-push gates on Phase 16 code
The mypy-strict and bandit hooks run only on push, so Phase 16 was never checked
against them during development. Fixes:
- derived.py: cast MSLP return to float (no-any-return)
- gain_drift.py: guard arg_max() None instead of int(None) (arg-type)
- ingest.py / alerts.py: parametrize dict return types; _row_to_dict over sqlite3.Row keys (type-arg, no-any-return, call-overload)
- notifier.py: httpx post content= (raw bytes body) not data= (also a correctness fix for ntfy)
- engine.py: contextlib.suppress instead of try/except/pass (bandit B110)

+++ observatory/weather/derived.py
@@ -44,7 +44,7 @@ def station_to_mslp(
     """
     T_K = temp_c + 273.15
     exponent = 5.2561
-    return pressure_hpa * (1.0 + (0.0065 * altitude_m) / T_K) ** exponent
+    return float(pressure_hpa * (1.0 + (0.0065 * altitude_m) / T_K) ** exponent)
 
 
 # ---------------------------------------------------------------------------
